fixed update-schema script
