+++ package.json
@@ -35,6 +35,7 @@
     "@types/styled-components": "5.1.4",
     "@types/styled-system": "5.1.10",
     "@types/supertest": "2.0.10",
+    "@types/webpack-node-externals": "^2.5.0",
     "@typescript-eslint/eslint-plugin": "4.8.1",
     "@typescript-eslint/parser": "4.8.1",
     "babel-jest": "26.6.3",
@@ -105,7 +106,7 @@
     "seed": "yarn b ./packages/server/scripts/seed.ts",
     "server": "yarn workspace @workshop/server start",
     "server:serve": "yarn workspace @workshop/server serve",
-    "update-schema": "yarn w ./packages/server/scripts/updateSchema.ts",
+    "update-schema": "yarn workspace @workshop/server update-schema",
     "web": "yarn workspace @workshop/web start",
     "w": "yarn b webpackx.ts",
     "start": "concurrently --kill-others \"yarn workspace @workshop/web start\" \"yarn workspace @workshop/server start\""
